requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
cloud-sql-python-connector==1.2.4
pg8000==1.29.6
SQLAlchemy==2.0.16
python-dotenv==0.20.0
gunicorn==20.1.0
psycopg2-binary==2.9.7